Fix major bug in downloader which made it unusuable for long lists
When there were long lists of codes to download with playlists in it the
downloader would be all over the place - taking forever to download and
having two downloads going on at once. That was due to the weird events
being fired by youtube-dl - 'end' in both cases, and 'next' for any vide
in a playlist. Refactored to handle these videos better, it seems to do
a much better job now