BOOSTNOTE begin to change architecture of script
There are now more objects and queues.
There is a global object 'notes' which is basically a database of
notes.
The YAML is loaded into an array 'rawIntructions', with minimal
cleanup. Each entry represents a url the user wishes to load.
From that a new array is derived, 'instructionsQueue'. This is a list of
async steps the script needs to carry out to get all the needed info.
For example, getting an image could be a step, getting the username from
another file could be another, etc. Each step updates the corresponding
data object within 'notes'.
Finally those notes are written to files.

TODO:
- so far only reddit was done
- old code is still knocking around
- everything is in one huge file, needs refactoring