01 boostnote refactor queues (#1)
* BOOSTNOTE begin to change architecture of script

There are now more objects and queues.
There is a global object 'notes' which is basically a database of
notes.
The YAML is loaded into an array 'rawIntructions', with minimal
cleanup. Each entry represents a url the user wishes to load.
From that a new array is derived, 'instructionsQueue'. This is a list of
async steps the script needs to carry out to get all the needed info.
For example, getting an image could be a step, getting the username from
another file could be another, etc. Each step updates the corresponding
data object within 'notes'.
Finally those notes are written to files.

TODO:
- so far only reddit was done
- old code is still knocking around
- everything is in one huge file, needs refactoring

* BOOSTNOTE refactor: add an InstructionsStore class and tests

- InstructionsStore deals with YAML data and replaces loadInstructions
  from lib
- There are now tests for InstructionsStore with some mock data

* BOOSTNOTE move files around and add linting

Moved all sources and tests inside src/ to make linting easier. Added
linting

* BOOSTNOTE add precommit hooks

* BOOSTNOTE add a Logger with tests

* BOOSTNOTE add info to Logger and implement in main script

* BOOSTNOTE add NotesStore and tests

- NotesStore has the data that will eventually be saved to Boostnote as
  notes. It has its own tests
- various notes mocks added inside __tests__/mockdata folder
- added jest-extend for more matchers

* BOOSTNOTE add eslint-jest and lint

* BOOSTNOTE InstructionsStore now writes to a log file

* BOOSTNOTE add foreach and update to InstructionsStore

* BOOSTNOTE add notes logging and ensure index has right loop

Notes has a log too, although the tests for it cannot run.
App produces following logs:
- xxx-log.yml : everything
- xxx-error.yml : errors
- xxx-unpublished.yml : no error, they simply didn't have a processor
- xxx-notes.yml : note
Inde.js now does nothing if running in dryRun, so that we can slowly
start processing type by type

* .gitignore

* BOOSTNOTE add a --clean option to clean logs folder

* BOOSTNOTE local tags override CLI tag for that card

* ANKI debug messages

* CPTN Improvements and extra functionality

- downloader now fires off post processing
- editor now has join and split

* PYTHON BOILERPLATE initial commit

* BOILERPLATE NODE initial commit

* ANKI add a quick script to add german words

* Package.json updates

* CPTN package.json updates